fix: tests editar perfil
